+++ NEWS.md
@@ -5,3 +5,8 @@
 # transmogR 1.1.1
 
 * Added `digestSalmon()`
+
+# transmogR 1.1.4
+
+* Changed default behaviour of `digestSalmon()` to exclude 'TPM' and 
+'effectiveLength' assays, which are now optional via the 'extra_assays' argument